+++ packages/core/examples/native/content.tsx
@@ -9,7 +9,6 @@ import { css, jsx } from '@emotion/react';
 import invariant from 'tiny-invariant';
 
 import Link from '@atlaskit/link';
-import { fg } from '@atlaskit/platform-feature-flags';
 import { Stack } from '@atlaskit/primitives/compiled';
 import { token } from '@atlaskit/tokens';
 
@@ -117,15 +116,10 @@ function DraggableAnchor() {
 			element,
 		});
 	}, []);
-	return fg('dst-a11y__replace-anchor-with-link__design-system-') ? (
+	return (
 		<Link href="#controlled" ref={ref}>
 			Anchor that is also a draggable()
 		</Link>
-	) : (
-		// eslint-disable-next-line @atlaskit/design-system/no-html-anchor
-		<a href="#controlled" ref={ref}>
-			Anchor that is also a draggable()
-		</a>
 	);
 }
 
@@ -143,15 +137,10 @@ function DraggableAnchorWithNewUrl() {
 			}),
 		});
 	}, []);
-	return fg('dst-a11y__replace-anchor-with-link__design-system-') ? (
+	return (
 		<Link href="#controlled" ref={ref}>
 			Link that is also a draggable() - with a new url
 		</Link>
-	) : (
-		// eslint-disable-next-line @atlaskit/design-system/no-html-anchor
-		<a href="#controlled" ref={ref}>
-			Link that is also a draggable() - with a new url
-		</a>
 	);
 }
 
@@ -179,15 +168,10 @@ function DraggableAnchorWithCustomPreview() {
 			},
 		});
 	}, []);
-	return fg('dst-a11y__replace-anchor-with-link__design-system-') ? (
+	return (
 		<Link href="#controlled" ref={ref}>
 			Link that is also a draggable() (custom preview)
 		</Link>
-	) : (
-		// eslint-disable-next-line @atlaskit/design-system/no-html-anchor
-		<a href="#controlled" ref={ref}>
-			Link that is also a draggable() (custom preview)
-		</a>
 	);
 }
 
@@ -240,26 +224,11 @@ export function Content() {
 			<div>
 				<ul>
 					<li>
-						{fg('dst-a11y__replace-anchor-with-link__design-system-') ? (
-							<Link href="#standalone">Standalone link</Link>
-						) : (
-							// eslint-disable-next-line @atlaskit/design-system/no-html-anchor
-							<a href="#standalone">Standalone link</a>
-						)}
+						<Link href="#standalone">Standalone link</Link>
 					</li>
 					<li>
-						{fg('dst-a11y__replace-anchor-with-link__design-system-') ? (
-							<Link href="#side-by">Side by</Link>
-						) : (
-							// eslint-disable-next-line @atlaskit/design-system/no-html-anchor
-							<a href="#side-by">Side by</a>
-						)}
-						{fg('dst-a11y__replace-anchor-with-link__design-system-') ? (
-							<Link href="#side-links">side links</Link>
-						) : (
-							// eslint-disable-next-line @atlaskit/design-system/no-html-anchor
-							<a href="#side-links">side links</a>
-						)}
+						<Link href="#side-by">Side by</Link>
+						<Link href="#side-links">side links</Link>
 					</li>
 					<li>
 						<DraggableAnchor />